requirements.txt
pandas
requests
transformers
torch
matplotlib
python-dotenv
tqdm